+++ actix-redis/src/session.rs
@@ -393,7 +393,7 @@ mod test {
     use super::*;
     use actix_session::Session;
     use actix_web::{
-        middleware, test, web,
+        middleware, web,
         web::{get, post, resource},
         App, HttpResponse, Result,
     };
@@ -493,7 +493,7 @@ mod test {
         //   - set-cookie actix-session will be in response (session cookie #3)
         //   - response should be: {"counter": 0, "user_id": None}
 
-        let srv = test::start(move || {
+        let srv = actix_test::start(|| {
             App::new()
                 .wrap(RedisSession::new(addr, &[0; 32]).cookie_name("test-session"))
                 .wrap(middleware::Logger::default())
@@ -686,7 +686,7 @@ mod test {
         //
         // Test that removing max_age results in a session-only cookie
         //
-        let srv = test::start(move || {
+        let srv = actix_test::start(|| {
             App::new()
                 .wrap(
                     RedisSession::new(addr, &[0; 32])