+++ actix-session/CHANGES.md
@@ -10,7 +10,7 @@
 * Update `actix-web` dependency to 4.0.0 beta.
 * Minimum supported Rust version (MSRV) is now 1.46.0.
 
-[#170]: https://github.com/actix/actix-extras/pull/???
+[#170]: https://github.com/actix/actix-extras/pull/170
 
 
 ## 0.4.1 - 2020-03-21